Assert root valid only in dev env
