+++ apps/extension/package.json
@@ -34,9 +34,9 @@
     "@radix-ui/react-icons": "^1.3.0",
     "@types/chrome": "0.0.246",
     "@types/firefox-webext-browser": "^111.0.2",
-    "@types/node": "^20.7.0",
-    "@types/react": "^18.2.22",
-    "@types/react-dom": "^18.2.7",
+    "@types/node": "^20.7.1",
+    "@types/react": "^18.2.23",
+    "@types/react-dom": "^18.2.8",
     "@typescript-eslint/eslint-plugin": "^6.7.3",
     "autoprefixer": "^10.4.16",
     "buffer": "^6.0.3",
