+++ apps/webapp/package.json
@@ -22,9 +22,9 @@
   },
   "devDependencies": {
     "@radix-ui/react-icons": "^1.3.0",
-    "@types/node": "^20.7.0",
-    "@types/react": "^18.2.22",
-    "@types/react-dom": "^18.2.7",
+    "@types/node": "^20.7.1",
+    "@types/react": "^18.2.23",
+    "@types/react-dom": "^18.2.8",
     "@typescript-eslint/eslint-plugin": "^6.7.3",
     "autoprefixer": "^10.4.16",
     "eslint-config-custom": "workspace:*",
