+++ packages/transport/package.json
@@ -14,11 +14,11 @@
     "@types/chrome": "0.0.246",
     "eslint": "^8.50.0",
     "eslint-config-custom": "workspace:*",
+    "react": "^18.2.0",
     "tsconfig": "workspace:*",
-    "typescript": "^5.2.2",
-    "react": "^18.2.0"
+    "typescript": "^5.2.2"
   },
   "devDependencies": {
-    "@types/react": "^18.2.22"
+    "@types/react": "^18.2.23"
   }
 }
