+++ packages/types/src/index.ts
@@ -5,3 +5,4 @@ export * from './indexed-db';
 export * from './utils';
 export * from './validation';
 export * from './lo-hi';
+export * from './environment';
