+++ packages/ui/package.json
@@ -21,9 +21,9 @@
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
-    "@types/node": "^20.7.0",
-    "@types/react": "^18.2.22",
-    "@types/react-dom": "^18.2.7",
+    "@types/node": "^20.7.1",
+    "@types/react": "^18.2.23",
+    "@types/react-dom": "^18.2.8",
     "autoprefixer": "^10.4.16",
     "eslint": "^8.50.0",
     "eslint-config-custom": "workspace:*",
